+++ src/main/java/org/summerboot/jexpress/boot/BackOffice.java
@@ -284,6 +284,8 @@ public class BackOffice extends BootConfig {
     @ConfigHeader(title = "4.2 Default Log4j2.xml Variables Naming")
     @Config(key = "naming.log4j2.xml.var.logId", defaultValue = "logId")
     private String log4j2LogId = "logId";
+    @Config(key = "naming.log4j2.xml.var.hitIndex", defaultValue = "hitIndex")
+    private String log4j2HitIndex = "hitIndex";
 
     @Config(key = "naming.log4j2.xml.var.logPath", defaultValue = "logPath")
     private String log4j2LogFilePath = "logPath";
@@ -522,6 +524,10 @@ public class BackOffice extends BootConfig {
 
     public String getLog4J2LogId() {
         return log4j2LogId;
+    }
+
+    public String getLog4j2HitIndex() {
+        return log4j2HitIndex;
     }
 
     public String getLog4j2LogFilePath() {