+++ src/main/java/org/summerboot/jexpress/integration/smtp/SMTPClientConfig.java
@@ -24,6 +24,7 @@ import org.summerboot.jexpress.boot.config.BootConfig;
 import org.summerboot.jexpress.boot.config.ConfigUtil;
 import org.summerboot.jexpress.boot.config.annotation.Config;
 import org.summerboot.jexpress.boot.config.annotation.ConfigHeader;
+import org.summerboot.jexpress.util.BackoffStrategy;
 
 import java.io.File;
 import java.util.HashSet;
@@ -110,6 +111,10 @@ public class SMTPClientConfig extends BootConfig {
     @Config(key = KEY_DEBOUCING_INTERVAL, defaultValue = "30", desc = "Alert message with the same subject will not be sent out within this minutes (default 30)")
     protected volatile int emailAlertDebouncingIntervalMinutes = 30;
 
+    @Config(key = "debouncing.emailalert.BackoffStrategy", defaultValue = "{\"strategy\":\"LINEAR\",\"initialInterval\":1800000,\"factor\":300000.0,\"maxInterval\":3600000,\"maxAttempts\":0,\"jitterFactor\":0.1}",
+            desc = "Backoff strategy, unit = millisecond), strategy = {LINEAR, EXPONENTIAL}, In exponential mode, factor is growth multiplier; in linear mode, factor is step. Unlimited attempts when maxAttempts <= 0")
+    protected volatile BackoffStrategy backoffStrategy = new BackoffStrategy(BackoffStrategy.Strategy.LINEAR, 1_800_000, 300000, 3_600_000, 0, 0.1);
+
     //3. mail session for Json display only
     protected Properties mailSessionProp;
 
@@ -181,5 +186,9 @@ public class SMTPClientConfig extends BootConfig {
 
     public int getEmailAlertDebouncingIntervalMinutes() {
         return emailAlertDebouncingIntervalMinutes;
+    }
+
+    public BackoffStrategy getBackoffStrategy() {
+        return backoffStrategy;
     }
 }